[AIP-21] Move hooks.sqlite_hook providers.sqlite.hooks.sqlite
