+++ airflow/operators/sqlite_operator.py
@@ -18,8 +18,8 @@
 # under the License.
 from typing import Iterable, Mapping, Optional, Union
 
-from airflow.hooks.sqlite_hook import SqliteHook
 from airflow.models import BaseOperator
+from airflow.providers.sqlite.hooks.sqlite import SqliteHook
 from airflow.utils.decorators import apply_defaults
