+++ airflow/providers/sqlite/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.sqlite.hooks.sqlite`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.sqlite.hooks.sqlite import SqliteHook  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.sqlite.hooks.sqlite`.",
-    DeprecationWarning, stacklevel=2
-)
