+++ docs/operators-and-hooks-ref.rst
@@ -1135,7 +1135,7 @@ These integrations allow you to perform various operations using various softwar
 
    * - `SQLite <https://www.sqlite.org/index.html>`__
      -
-     - :mod:`airflow.hooks.sqlite_hook`
+     - :mod:`airflow.providers.sqlite.hooks.sqlite`
      - :mod:`airflow.operators.sqlite_operator`
      -
